`gfx` D3D12 shader objects rewrite.
This change reimplements D3D12 shader objects in the `gfx` layer,
so that it is no longer layered over the `DescriptorSet`
abstraction. The new implementation also takes advantage of
D3D12's flexibility to map a `space` and `register` binding to an
arbitrary descriptor address to facilitate a more efficient
binding algorithm.

In this implementation, each shader object
(representing an entry-point, a constant buffer or a parameter
block) maintains one descriptor table set (one table for each
descriptor heap type) that is used to hold all bindings for the
object, excluding bindings of sub-objects. For example, if the
global scope defines the following parameters:
```
struct CB
{
    Texture2D texture;
    SamplerState s;
};
ConstantBuffer<CB> cb;
RWStructuredBuffer<int> bufferOut;
```
Then we will use two shader objects to hold all parameters:
```
ShaderObject_0 (root):
    ResourceDescriptorTable(slotCount = 1)
        DescriptorRangeType: UAV
        DescriptorCount: 1
ShaderObject_1 (cb):
    ResourceDescriptorTable(slotCount = 1)
        DescriptorRangeType: SRV
        DescriptorCount: 1
    SamplerDescriptorTable(slotCount = 1)
        DescriptorRangeType: Sampler
        DescriptorCount: 1
```
We will create the following root signature:
```
RootParam[0]:
    DescriptorTable
        Range[0]: UAV, Count = 1
        Range[1]: CBV, Count = 1
RootParam[1]:
    DescriptorTable
        Range[0]: Sampler, Count = 1
```

Note that in this root signature, the descriptor for the
constant buffer `cb` appears *after* the UAV for `bufferOut`,
despite that `cb` is defined before `bufferOut` in the shader
code. This is intentional: by laying out the descriptors in
this order, we can simply copy the descriptor tables from all
shader objects to form the final GPU descriptor table.
In this example, when we are generating the GPU descriptor
from the hierarchy of shader objects, we simply append the
descriptor tables for the global scope and for `cb`.

By doing so, the user's `space` and `register` bindings will
only has an effect on the root signature's descriptor range
descriptions, and will not affect the descriptor table layout.
This allows the `gfx` layer to always operate the in most
efficient way.